fixed code.

+++ 025-DOM-selectors/main.js
@@ -35,9 +35,9 @@ function isFormValid(miles, gallons, price) {
 FORM.addEventListener("submit", (e) => {
   e.preventDefault();
   //const errMsg = [];
-  const miles = parseInt(e.target.miles.value);
-  const gallons = parseInt(e.target.gallons.value);
-  const price = parseInt(e.target.price.value);
+  const miles = parseFloat(e.target.miles.value);
+  const gallons = parseFloat(e.target.gallons.value);
+  const price = parseFloat(e.target.price.value);
 
   const isValid = isFormValid(miles, gallons, price);
   if (isValid) {
